requirements.txt
backports.functools-lru-cache==1.6.4
beautifulsoup4==4.9.3
certifi==2021.10.8
chardet==4.0.0
lxml==4.9.1
requests==2.27.1
soupsieve==1.9.6
urllib3==1.26.12
grequests~=0.6.0
pillow==6.2.2
pygame=1.9.2